Add ena.samples virtual table — INSERT support (Phase 5)
ATTACH '' AS ena (TYPE ENA, SECRET 'name') now exposes ena.samples
alongside ena.projects. INSERT INTO ena.samples ... RETURNING
alias, ers_accession, samea_accession round-trips through the
Webin V2 envelope, returning the BioSample-first accessions users
need for grant reporting and consortium coordination before any
sequencing data exists. This is MILESTONE 1 of the submission plan.

End-to-end SQL:
    CREATE SECRET my_ena (TYPE ENA, USER 'Webin-X', PASSWORD_ENV 'PW');
    ATTACH '' AS ena (TYPE ENA, SECRET 'my_ena');
    INSERT INTO ena.projects (alias, title, project_type)
      VALUES ('p1', 'Demo', 'METAGENOMIC')
      RETURNING prjeb_accession, erp_accession;
    INSERT INTO ena.samples (alias, taxon_id, checklist, attributes)
      VALUES ('s1', 408170, 'ERC000015',
              MAP {'collection date': '2026-04-01', ...})
      RETURNING ers_accession, samea_accession;

Pure-data layer (src/ena_samples_insert.{cpp,hpp}):
- SubmitSampleInsertOutcome builds the V2 envelope, POSTs via the
  injected functor, parses the receipt, and joins receipt SAMPLE
  entries to the request by alias
- Logical failures (server success=false, parse errors, missing
  alias) populate outcome.success=false + error_messages while
  preserving raw_receipt — same contract as projects
- ext_id/@type matched case-insensitively against "biosample" so
  receipts with different casings ("BioSample", "BIOSAMPLE") still
  resolve the SAMEA — caught in code review

PhysicalOperator (src/ena_samples_insert_op.{cpp,hpp}):
- Sink buffers chunks, Finalize drains them, BuildSamplesFromBuffer
  maps DuckDB column values to SampleSpec honoring column_index_map
- ExtractAttributesMap translates MAP(VARCHAR,VARCHAR) -> attribute
  list pairs for the envelope (NULL keys / mismatched struct shape
  rejected with a clear error)
- AppendReturningRows emits the full 9-column samples layout; the
  attributes column round-trips as NULL (user-supplied MAP is
  preserved verbatim in submission_log.request_payload, same as
  projects.attributes in Phase 4)

Shared infrastructure:
- src/ena_insert_common.{cpp,hpp} — extracted from the projects
  operator: ResolveENACredentials, ResolveInputColumn, value
  coercion helpers, GenerateSubmissionId, parameterized
  RecordSubmissionLog (takes a SubmissionLogPayload with
  object_type / action / etc.)
- src/include/ena_post_fn.hpp — tiny dedicated header for the
  ENAPostFn typedef so per-table insert headers no longer have to
  cross-include each other
- src/include/ena_receipt_parser.hpp adds inline EqualsIgnoreCase
  helper (pure-data layer cannot link against StringUtil)

Schema additions:
- SampleSpec re-introduces description and scientific_name fields
  (Phase 3 had them as scaffolded-then-removed dead code) — the
  envelope builder emits both when set
- ena.submission_log.duration_ms changed from INTEGER to BIGINT
  (and the row struct from int32 to int64); int32 ms caps at ~35
  minutes which a slow multi-hundred-sample submission could
  plausibly exceed

Catalog dispatcher: ENACatalog::PlanInsert now switches on table
kind, routing to ENAProjectsInsert OR ENASamplesInsert. Other
tables still throw "not implemented" until their phase lands.

Tests:
- test/cpp/test_ena_samples_insert.cpp: 7 cases / 33 assertions
  covering single-row, multi-sample, success=false, empty input,
  missing receipt entry, HOLD round-trip, scientific_name +
  description envelope shape
- test/sql/ena_samples_insert_mock.test: round-trip via the local
  Python mock — single, multi, FAIL alias, taxon_id<=0 rejection,
  empty-alias rejection (require httpfs, skipped without)
- test/sql/ena_samples_insert_live.test: same shape vs
  wwwdev.ebi.ac.uk, gated on ENA_WEBIN_TEST_USER/PASSWORD/AVAILABLE
- test/scripts/ena_webin_mock.py: SAMPLE accession layout fixed —
  primary is now ERS, EXT_ID is SAMEA (was reversed; caught by
  manual round-trip during the GREEN step)

Tests delta: +7 C++ cases (764 total), +33 assertions (6696 total).
SQL tier: +2 files (skipped without httpfs/creds). 0 regressions.

Phase 5 also picked up follow-up improvements caught by code review:
- Per-row ResolveInputColumn calls hoisted out of the Sink-side row
  loop in both ena_projects_insert_op.cpp and ena_samples_insert_op
  (column_index_map is per-statement, not per-row)
- Reports API SELECT path deferred to a future phase (BioSample-first
  workflow doesn't need read-back; design doc §7.4 will be wired
  when there's a consumer)
- Extract a templated Finalize body / CRTP Sink-Source base before
  Phase 7 lands experiments/runs/analyses operators (recorded in
  the plan's Phase 5 done log)

+++ src/include/ena_storage.hpp
@@ -132,7 +132,7 @@ struct ENASubmissionLogRow {
 	string request_payload;
 	string receipt;
 	vector<string> error_messages;
-	int32_t duration_ms;
+	int64_t duration_ms;
 };
 
 class ENASubmissionLog {
